[color] subdued color for orgmode
This commit tweaks colors theme for orgmode to minimize the color distracting.
Find appropriate light colors and interpolate between gray90 (the default text
color) and the light colors to produce the headline colors.

The ideal outcome would be minimize the contrast among different elements in the
text but to maintain certain eligibility.