Make sure directory paths are treated as directory.
A directory is also a file, but its name as a directory is different
from its name as a file.  For example, in Unix-like systems, a trailing
slash '/' is expected for a directory name.

In Emacs, some functions behave differently depending on if it is a
directory or file name.